Updated plot_r and plot_p for correlation ensembles
Created 2 new functions, plot_corr.ens() and plot_pvals.ens() in
plotting.geoChronR.R, because the job was too big for plot_hist.ens().
I kept the latter for back-compatibility.

Also updated cor.ens() in correlation.regression.functions.R to make
the correct function calls. Have not tinkered with regression at all
but suggest that similar methods could be applied there.

Note: everything currently assumes that the test level is 5%, but this
cna be easily relaxed with a little tinkering and a cunning use of
bquote().

Also, the default quantiles for everything are 2.5% 25% 50% 75% 97.5%,
because the normal assumption is generally too restrictive with
distributions like these.